Rewrite most of the script...
The original script did not correctly dump the root directory object, and it did not correctly enumerate an object directory to begin with. I was not sure if the point was to dump the contents of "\ObjectTypes", as to do this, the script would have to take a very different approach, or to simply dump all of the objects under "\" (ObpRootDirectoryObject). I assumed the intent was the latter, as there is already a script in the repository to dump all object types, and rewrote the script to do so.